Complete Burn feature, misc. cleanup
